tdxsec: introduce manifest files for meta-toradex-security
This commit introduces manifest files aimed at automating the build and
testing of features provided by the meta-toradex-security layer.

Since we want to use metadata provided meta-toradex-distro,
meta-toradex-demos, and others, we build on top of tdxref manifests,
avoiding the need for content duplication.

As there are no plans to establish a formal release cadence for
meta-toradex-security at this time, a manifest with a fixed hash
(default.xml) is not included.

Signed-off-by: Sergio Prado <sergio.prado@e-labworks.com>