torizon: integration: remove comment about update external layers hashes
Commit bf26d48562af added a comment about updating the external layers after
leaving the feature freeze. However, it was decided that doing that is not
aligned with the current strategy of bumping the external layers in the
manifest integration according to the result of the extint build.

Therefore, remove this comment, as this procedure is not aligned with how
extint is currently being used.

Related-to: TOR-3989

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>